refactor: Move the CassandraDBObjectStore into its own package
