+++ .vscode/settings.json
@@ -25,6 +25,9 @@
         "src/causal-tree-store-mongodb/**/*.d.ts": {
             "when": "$(basename).ts"
         },
+        "src/causal-tree-store-cassandradb/**/*.d.ts": {
+            "when": "$(basename).ts"
+        },
         "src/causal-tree-store-browser/**/*.d.ts": {
             "when": "$(basename).ts"
         },
