+++ src/aux-server/package.json
@@ -51,6 +51,7 @@
         "@casual-simulation/causal-tree-server": "^1.0.26",
         "@casual-simulation/causal-tree-server-socketio": "^1.0.26",
         "@casual-simulation/causal-tree-store-mongodb": "^1.0.26",
+        "@casual-simulation/causal-tree-store-cassandradb": "^1.0.26",
         "@casual-simulation/causal-trees": "^1.0.26",
         "@casual-simulation/crypto": "^1.0.24",
         "@casual-simulation/crypto-browser": "^1.0.24",
