+++ src/aux-server/server/config.ts
@@ -1,4 +1,5 @@
 import { WebConfig } from '../shared/WebConfig';
+import { CassandraDBCausalReposConfig } from '@casual-simulation/causal-tree-store-cassandradb';
 import redis from 'redis';
 
 export const DRIVES_URL = '/drives';
@@ -101,34 +102,6 @@ export interface CausalReposServerConfig {
 
 export interface MongoDBCaualReposConfig {
     dbName: string;
-}
-
-export interface CassandraDBCausalReposConfig {
-    /**
-     * The keyspace that should be used for the tables.
-     */
-    keyspace: string;
-
-    /**
-     * The replication parameters that should be used for the keyspace.
-     * If null, then creating the keyspace will be skipped.
-     */
-    replication: CassandraDBReplication | null;
-}
-
-export type CassandraDBReplication =
-    | CassandraDBSimpleReplication
-    | CassandraDBNetworkTopologyReplication;
-
-export interface CassandraDBSimpleReplication {
-    class: 'SimpleStrategy';
-    replicationFactor: number;
-}
-
-export interface CassandraDBNetworkTopologyReplication {
-    class: 'NetworkTopologyStrategy';
-    replicationFactor: number;
-    dataCenters: { [key: string]: any };
 }
 
 /**
