+++ src/causal-tree-store-cassandradb/.gitignore
@@ -1,0 +1,6 @@
+*.js
+*.js.map
+*.d.ts
+!gulpfile.js
+!typings/**/*
+docs
