+++ src/causal-tree-store-cassandradb/README.md
@@ -1,0 +1,5 @@
+# Causal Tree Store CassandraDB
+
+[![npm (scoped)](https://img.shields.io/npm/v/@casual-simulation/causal-tree-store-cassandradb.svg)](https://www.npmjs.com/package/@casual-simulation/causal-tree-store-cassandradb)
+
+A causal repo store that interfaces with CassandraDB.
