+++ src/causal-tree-store-cassandradb/index.ts
@@ -1,0 +1,2 @@
+export * from './CassandraDBCausalReposConfig';
+export * from './CassandraDBObjectStore';
